Fix corner case when /v2/apps returns apps across all spaces.
There's a corner case given a developer has access to multiple spaces
where there are apps with the same name (despite having different
routes). Prior to this patch, it would `DoesAppExist` would return
false because it would see more than one app thus causing the script to
deploy as a first time deploy which would really cause downtime if it's
pushing to an existing app in the space. This patch will search for the
apps and correlate between app and it's space and count them.